Move generic utils in lib/utils
