+++ src/bin/cli.ts
@@ -8,7 +8,7 @@ import ncu from '../index'
 import { chalkInit, getChalk } from '../lib/chalk'
 // async global contexts are only available in esm modules -> function
 import getNcuRc from '../lib/getNcuRc'
-import { pickBy } from '../lib/pick'
+import { pickBy } from '../lib/utils/pick'
 import { type RunOptions } from '../types/RunOptions'
 
 const optionVersionDescription = 'Output the version number of npm-check-updates.'
