+++ src/cli-options.ts
@@ -2,8 +2,8 @@ import path from 'path'
 import { defaultCacheFile } from './lib/cache'
 import chalk from './lib/chalk'
 import parseCooldown from './lib/parseCooldown'
-import { sortBy } from './lib/sortBy'
 import table from './lib/table'
+import { sortBy } from './lib/utils/sortBy'
 import type CLIOption from './types/CLIOption'
 import type ExtendedHelp from './types/ExtendedHelp'
 import { type Index } from './types/IndexType'
