+++ src/lib/getIgnoredUpgradesDueToEnginesNode.ts
@@ -6,8 +6,8 @@ import { type Options } from '../types/Options'
 import { type Version } from '../types/Version'
 import { type VersionSpec } from '../types/VersionSpec'
 import getEnginesNodeFromRegistry from './getEnginesNodeFromRegistry'
-import keyValueBy from './keyValueBy'
 import upgradePackageDefinitions from './upgradePackageDefinitions'
+import keyValueBy from './utils/keyValueBy'
 
 /** Checks if package.json min node version satisfies given package engine.node spec */
 const satisfiesNodeEngine = (enginesNode: Maybe<VersionSpec>, optionsEnginesNodeMinVersion: Version) =>
