+++ src/lib/utils/escapeRegExp.ts
