+++ src/lib/utils/pick.ts
