+++ src/package-managers/npm.ts
@@ -9,10 +9,10 @@ import nodeSemver from 'semver'
 import { parseRange } from 'semver-utils'
 import untildify from 'untildify'
 import pkg from '../../package.json'
-import { keyValueBy } from '../lib/keyValueBy'
 import libnpmconfig from '../lib/libnpmconfig'
 import { print, printSorted } from '../lib/logging'
 import spawnCommand from '../lib/spawnCommand'
+import { keyValueBy } from '../lib/utils/keyValueBy'
 import * as versionUtil from '../lib/version-util'
 import type { GetVersion } from '../types/GetVersion'
 import type { Index } from '../types/IndexType'
