+++ src/package-managers/pnpm.ts
@@ -5,9 +5,9 @@ import ini from 'ini'
 import os from 'os'
 import path from 'path'
 import { parse as parseYaml } from 'yaml'
-import keyValueBy from '../lib/keyValueBy'
 import { print } from '../lib/logging'
 import spawnCommand from '../lib/spawnCommand'
+import keyValueBy from '../lib/utils/keyValueBy'
 import { type GetVersion } from '../types/GetVersion'
 import { type Index } from '../types/IndexType'
 import { type NpmConfig } from '../types/NpmConfig'
