+++ src/package-managers/yarn.ts
@@ -4,12 +4,12 @@ import jsonlines from 'jsonlines'
 import os from 'os'
 import path from 'path'
 import { parse as parseYaml } from 'yaml'
-import exists from '../lib/exists'
 import findLockfile from '../lib/findLockfile'
-import { keyValueBy } from '../lib/keyValueBy'
 import { print } from '../lib/logging'
 import parseCooldown from '../lib/parseCooldown'
 import spawnCommand from '../lib/spawnCommand'
+import exists from '../lib/utils/exists'
+import { keyValueBy } from '../lib/utils/keyValueBy'
 import { type GetVersion } from '../types/GetVersion'
 import { type Index } from '../types/IndexType'
 import { type NpmConfig } from '../types/NpmConfig'
