+++ test/install.test.ts
@@ -8,7 +8,7 @@ import { stripVTControlCharacters as stripAnsi } from 'node:util'
 import os from 'os'
 import path from 'path'
 import spawn from 'spawn-please'
-import exists from '../src/lib/exists'
+import exists from '../src/lib/utils/exists'
 import chaiSetup from './helpers/chaiSetup'
 import removeDir from './helpers/removeDir'
 import stubVersions from './helpers/stubVersions'
